Unify emoji count and emoji information to a cog.
